+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- The ``codename`` field and related stuff has been removed from ``DatabaseMapping``.
+
 ### Fixed
 
 ### Security